Refactor run.py
Broke functions into modules
